fix: surveyEditor issues
